refactor: move input replace to executor
