+++ src/oocana/__init__.py
@@ -1,3 +1,2 @@
-from .func import *  # noqa: F403
 from .data import *  # noqa: F403
 from .context import *  # noqa: F403
